Don't show send notifications box if config doesn't enable notifications
